requirements.txt
Babel==2.7.0
Django==2.2.3
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==2.0
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.5
factory-boy==2.12.0
Faker==1.0.7
phonenumbers==8.10.14
pika==1.0.1
Pillow==6.1.0
purl==1.5
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
ujson==1.35
